rename binary to starsd
