+++ cmd/starsd/cmd/genaccounts.go
