+++ cmd/starsd/cmd/init.go
