+++ cmd/starsd/cmd/testnet.go
@@ -61,7 +61,7 @@ func testnetCmd(mbm module.BasicManager, genBalIterator banktypes.GenesisBalance
 necessary files (private validator, genesis, config, etc.).
 Note, strict routability for addresses is turned off in the config file.
 Example:
-	staked testnet --v 4 --output-dir ./output --starting-ip-address 192.168.10.2
+	starsd testnet --v 4 --output-dir ./output --starting-ip-address 192.168.10.2
 	`,
 		RunE: func(cmd *cobra.Command, _ []string) error {
 			clientCtx := client.GetClientContextFromCmd(cmd)
@@ -117,7 +117,7 @@ Example:
 	cmd.Flags().StringP(flagOutputDir, "o", "./mytestnet", "Directory to store initialization data for the testnet")
 	cmd.Flags().String(flagNodeDirPrefix, "node",
 		"Prefix the directory name for each node with (node results in node0, node1, ...)")
-	cmd.Flags().String(flagNodeDaemonHome, "staked", "Home directory of the node's daemon configuration")
+	cmd.Flags().String(flagNodeDaemonHome, "starsd", "Home directory of the node's daemon configuration")
 	cmd.Flags().String(
 		flagStartingIPAddress,
 		"192.168.0.1",
@@ -514,7 +514,7 @@ services:{{range $node := .Nodes }}
 			- {{ $node.APIPort}}:1317
 			- {{ $node.GRPCPort}}:9090
 		volumes:
-			- ./{{$node.Name}}/staked:/data/.staked/
+			- ./{{$node.Name}}/starsd:/data/.starsd/
 {{end}}
 `
