+++ contrib/build.sh
@@ -1,1 +1,1 @@
-GOARCH=amd64 GOOS=linux go build -o bin/staked github.com/public-awesome/stargaze/cmd/staked
+GOARCH=amd64 GOOS=linux go build -o bin/starsd github.com/public-awesome/stargaze/cmd/starsd
