+++ docker/Dockerfile.test
@@ -2,7 +2,7 @@ FROM alpine:3.12
 RUN apk add -U --no-cache ca-certificates
 
 WORKDIR /root
-COPY ./bin/staked /usr/bin/staked
+COPY ./bin/starsd /usr/bin/starsd
 COPY ./docker/test-entry-point.sh ./entry-point.sh
 EXPOSE 26657
